tests: sectionconfig: add case for unused optional default properties
Add a PVE::SectionConfig test case for plugin systems that declare an
optional default property (in `private()->{propertyList}`) but don't
use it anywhere, i.e. not declaring it in any `options()` method.

In unified mode, such properties are included in the create schema,
but *not* in the update schema.

In isolated mode, such properties are included in both schemas.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-6-m.carrara@proxmox.com