tests: sectionconfig: add case for opt. default prop being req. once
Add a PVE::SectionConfig test case for plugin systems that declare an
optional default property (in `private()->{propertyList}`) and require
it on at least one, but not all child plugins.

In other words, only one child plugin uses the optional default
property in its `options()` method, and does so with `optional => 0`.

In unified mode, such properties are marked as *optional* in both
create and update schema.

In isolated mode, such properties are only available for the plugins
that use them in `options()`, and for those plugins, they are
*optional* despite being marked as required.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-7-m.carrara@proxmox.com