tests: sectionconfig: add case for opt. default prop requirde by all
Add a PVE::SectionConfig test case for plugin systems that declare an
optional default property (in `private()->{propertyList}`) and mark it
as required on all child plugins.

In other words, all child plugins use the optional default property in
their `options()` methods, and do so with `optional => 0`.

In unified mode, such properties are marked as *required* in both
create and update schema.

In isolated mode, such properties are marked as *optional* in both
create and update schema.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-8-m.carrara@proxmox.com