tests: sectionconfig: add case for req. default prop being req. once
Add a PVE::SectionConfig test case for plugin systems that declare a
required default property (in `private()->{propertyList}`) and require
it on at least one, but not all child plugins.

In other words, one child plugin uses the required default property in
its `options()` method, and does so with `optional => 0`.

In unified mode, such properties are marked as *required* in both
create and update schema.

In isolated mode, such properties are only available for the plugins
that use them in `options()`, and for those plugins they are
*optional* in both create and update schema, despite being marked as
required.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-10-m.carrara@proxmox.com