tests: sectionconfig: add case for required default props opt. for all
Add a PVE::SectionConfig test case for plugin systems that declare a
default property as required (in `private()->{propertyList}`), but
mark it as optional in all child plugins (i.e. in their `options()`
methods).

In unified mode, such properties remain marked as *required* in both
create and update schema.

In isolated mode, such properties are marked as *optional* in both
create and update schema.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-12-m.carrara@proxmox.com