tests: sectionconfig: add isolated mode test structure
Similar to `schema_comparison_test.pl`, add `schema_isolated_test.pl`
for tests specific to PVE::SectionConfig's property isolation mode.

The purpose of these tests is to document the existing quirks of the
isolated mode of PVE::SectionConfig as well as to ensure that we catch
any regressions, should we need to make any changes to the inner
machinery of PVE::SectionConfig in the future.

This test script works similar to `schema_comparison_test.pl`, expect
that its cases are (obviously) not tested against their counterparts
in unified mode, as that simply doesn't apply for them.

Add a simple basic test case for child plugins that define identical
properties under isolated mode.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-13-m.carrara@proxmox.com