tests: sectionconfig: add case for same prop. w/ different optionality
Add a PVE::SectionConfig isolated mode test case where a property with
the same name, but differing optionality is defined on two plugins.
That is, the name is the same, but one has `optional => 1`, whereas
the other has `optional => 0`.

In both create and update schema, both properties are marked as
optional, because properties added by child plugins are always marked
as optional.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-15-m.carrara@proxmox.com