tests: sectionconfig: add case for differing opt. default prop uses
Add a PVE::SectionConfig isolated mode test case where an optional
default property (i.e. it's defined in `private()->{propertyList}`) is
marked as required by one plugin, but as optional by the other.

In both create and update schema, such a property is marked as
optional for plugins that use it, even for plugins that mark it as
required.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-16-m.carrara@proxmox.com