tests: sectionconfig: add case for differing req. default prop uses
Add a PVE::SectionConfig isolated mode test case where a required
default property (i.e. it's defined in `private()->{propertyList}`) is
marked as required by one plugin, but as optional by the other.

In the create schema, such a property uses a `oneOf` schema to differ
between the cases of both plugins. In other words, if a plugin marked
it as optional, it is marked as optional in the create schema too, and
vice versa.

In the update schema, such a property is marked as optional for both
plugins.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-17-m.carrara@proxmox.com