sectionconfig: reword docs regarding property usage in isolated mode
In the MODES section, clarify that declaring a plugin's own properties
in `options()` is not necessary in isolated mode for visibility's
sake. The phrase "automatically added to its schema" might otherwise
not be completely clear to lesser experienced users.

Also reword a similar paragraph in the docstring for the
`properties()` method, making it a little less wordy overall.
Additionally, that paragraph mentions that a local property may be
marked as required in `options()`, which is incorrect, so get rid of
that passage of text too.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-19-m.carrara@proxmox.com