sectionconfig: extend / correct docstring of `private()` method
At the end of the docstring, note that ..

* one really should only declare default properties as required in
  very special cases, such as using a property as an identifier

* and that only attributes specific to PVE::JSONSchema should be
  declared on default properties

The latter case in particular was mentioned off-list a while ago;
every default property must be defined through a valid JSONSchema, any
other attributes really shouldn't be added.

Also correct the example in this docstring along the way, correctly
quoting the key and adding `optional => 1` in order not to suggest
anything that is advised against in the new paragraph further below.

Signed-off-by: Max R. Carrara <m.carrara@proxmox.com>
Link: https://lore.proxmox.com/all/20251219194511.840583-20-m.carrara@proxmox.com