fix #7193: allow vlan-interfaces as physical bridge ports
The bug addressed in [0] is also affecting current PVE 8.4
installations.

[0]: https://lore.proxmox.com/pve-devel/20251230160730.641868-1-s.ivanov@proxmox.com/T/#u

This patch backports the changes for 9 to PVE 8 (backport consists of
moving the code to PVE/Network.pm as PVE/IPRoute2.pm was introduced in
PVE 9 and adapting to the lack of `use v5.36;` in PVE8).

short summary of the cause of the issue: commit 057f62f ("fix #7118:
fix bridge port detection when plugging netdev with vlan") did not
address that ports in the form of eno1.1234 (vlan 1234 on eno1) were
allowed as bridge-ports.

This is a semantic backport of commit ca27acb ("fix #7193: allow
vlan-interfaces as physical bridge ports") from master.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20251230162415.665391-1-s.ivanov@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>